Fix slideshow

+++ foostart/package-slideshow/Controllers/Admin/StyleAdminController.php
@@ -171,10 +171,11 @@ class StyleAdminController extends FooController
                 }
 
                 // add new item
-            } else if (!$this->obj_validator->isExistingView($params)){
-
+            } else {
                 //save to file
-                file_put_contents($params['view_style_path'] . '/' . $params['style_view_file'] . '.blade.php', $params['style_view_content']);
+                if (!empty($params['style_view_file']) && !$this->obj_validator->isExistingView($params)) {
+                    file_put_contents($params['view_style_path'] . '/' . $params['style_view_file'] . '.blade.php', $params['style_view_content']);
+                }
                 //insert
                 $item = $this->obj_item->insertItem($params);
 
@@ -240,5 +241,50 @@ class StyleAdminController extends FooController
         return Redirect::route($this->root_router . '.list')
             ->withMessage(trans($this->plang_admin . '.actions.delete-error'));
     }
+
+
+    /**
+     * Edit existing item by {id} parameters OR
+     * Add new item
+     * @return view edit page
+     * @date 26/12/2017
+     */
+    public function copy(Request $request)
+    {
+        $item = NULL;
+        $categories = NULL;
+
+        $params = $request->all();
+        $params['id'] = $request->get('cid', NULL);
+
+        if (!empty($params['id'])) {
+
+            $item = $this->obj_item->selectItem($params, FALSE);
+
+            if (empty($item)) {
+                return Redirect::route($this->root_router . '.list')
+                    ->withMessage(trans($this->plang_admin . '.actions.edit-error'));
+            }
+
+            $item->id = NULL;
+        }
+
+        //get categories by context
+        $context = $this->obj_item->getContext($this->category_ref_name);
+        if ($context) {
+            $params['context_id'] = $context->context_id;
+            $categories = $this->obj_category->pluckSelect($params);
+        }
+
+        // display view
+        $this->data_view = array_merge($this->data_view, array(
+            'item' => $item,
+            'categories' => $categories,
+            'request' => $request,
+            'context' => $context,
+        ));
+        return view($this->page_views['admin']['edit'], $this->data_view);
+    }
+
 
 }
